refactor(cypress): keep command a JS file, to avoid cypress types conflicts
